requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0

httpx>=0.27.0
trafilatura>=1.8.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
tenacity>=8.2.0
lxml_html_clean>=0.4.0
openai>=1.12.0
chromadb>=0.5.0